requirements.txt
arrow==1.4.0
asarPy==1.0.1
asgiref==3.11.1
binaryornot==0.4.4
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
cookiecutter==2.6.0
dj-database-url==3.1.0
Django==6.0.2
django-cloudinary-storage==0.3.0
djangorestframework==3.16.1
dotenv==0.9.9
gunicorn==25.0.3
honcho==2.0.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==26.0
pillow==12.1.0
psycopg==3.3.2
psycopg-binary==3.3.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
rich==14.3.2
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
tzdata==2025.3
urllib3==2.6.3
whitenoise==6.11.0
